requirements.txt
git+https://github.com/SENERGY-Platform/analytics-operator-lib-python@v1.1.29

pandas<2
kazoo<3
astral
pytz
scipy
tqdm
similaritymeasures
pyarrow
scikit-learn
kneed[plot]
python-dotenv
numpy
